requirements.txt
aioredis==1.3.1
asgiref==3.7.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
autopep8==2.0.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
constantly==15.1.0
crispy-bootstrap4==2022.1
cryptography==36.0.1
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2.1
django-allauth==0.48.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.1.0
django-devserver==0.8.0
django-multiupload==0.6.1
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==19.2.0
gprof2dot==2022.7.29
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
imageio==2.31.1
imageio-ffmpeg==0.4.8
incremental==21.3.0
inflection==0.5.1
iniconfig==2.0.0
moviepy==1.0.3
msgpack==1.0.3
numpy==1.25.1
oauthlib==3.2.0
packaging==21.3
Pillow==10.0.0
pluggy==1.2.0
priority==1.3.0
proglog==0.1.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.1
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tabulate==0.9.0
tomli==2.0.1
tqdm==4.65.0
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
zope.interface==5.4.0
